+++ app/src/main/java/com/eddyizm/tempus/ui/login/LoginActivity.kt
@@ -1,7 +1,9 @@
 package com.eddyizm.tempus.ui.login
 
 import android.os.Bundle
+import android.view.MenuItem
 import android.view.View
+import android.widget.Toast
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
@@ -21,6 +23,7 @@ private const val SERVERS_FRAGMENT: Int     = 3
 class LoginActivity : AppCompatActivity() {
     private lateinit var binding: ActivityLoginBinding
     private var hideTabLayout: Boolean = false
+    private var hideTopAppBar: Boolean = true
     private var selectedFragment: Int = GREETER_FRAGMENT
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -33,11 +36,14 @@ class LoginActivity : AppCompatActivity() {
 
         initIntentHandler()
         initEdgeToEdge()
+        initTopAppBarLayout()
         initTabLayout()
     }
 
     private fun initIntentHandler() {
         hideTabLayout    = intent.getBooleanExtra("HIDE_TAB_LAYOUT", false)
+        hideTopAppBar    = intent.getBooleanExtra("HIDE_TOPAPPBAR_LAYOUT", true)
+
         selectedFragment = intent.getIntExtra("SELECT_FRAGMENT", GREETER_FRAGMENT)
     }
 
@@ -47,6 +53,14 @@ class LoginActivity : AppCompatActivity() {
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
             insets
         }
+    }
+
+    private fun initTopAppBarLayout() {
+        if (hideTopAppBar) {
+            binding.topAppBar.visibility = View.GONE
+            return
+        }
+        binding.topAppBar.setOnClickListener { finish() }
     }
 
     private fun initTabLayout() {
@@ -75,18 +89,30 @@ class LoginActivity : AppCompatActivity() {
 
         override fun createFragment(position: Int): Fragment {
             return when (position) {
-                GREETER_FRAGMENT -> LoginGreeterFragment.newInstance(
-                    singlePageMode = hideTabLayout
-                )
-                PERMISSIONS_FRAGMENT -> LoginPermissionFragment.newInstance(
-                    singlePageMode = hideTabLayout
-                )
-                THEMES_FRAGMENT -> LoginThemeFragment.newInstance(
-                    singlePageMode = hideTabLayout
-                )
-                SERVERS_FRAGMENT -> LoginServerFragment.newInstance(
-                    singlePageMode = hideTabLayout
-                )
+                GREETER_FRAGMENT -> {
+                    binding.topAppBar.title = getString(R.string.la_appbar_greeter_title)
+                    LoginGreeterFragment.newInstance(
+                        singlePageMode = hideTabLayout
+                    )
+                }
+                PERMISSIONS_FRAGMENT -> {
+                    binding.topAppBar.title = getString(R.string.la_appbar_permission_title)
+                    LoginPermissionFragment.newInstance(
+                        singlePageMode = hideTabLayout
+                    )
+                }
+                THEMES_FRAGMENT -> {
+                    binding.topAppBar.title = getString(R.string.la_appbar_themes_title)
+                    LoginThemeFragment.newInstance(
+                        singlePageMode = hideTabLayout
+                    )
+                }
+                SERVERS_FRAGMENT -> {
+                    binding.topAppBar.title = getString(R.string.la_appbar_server_title)
+                    LoginServerFragment.newInstance(
+                        singlePageMode = hideTabLayout
+                    )
+                }
                 else -> LoginGreeterFragment.newInstance(
                     singlePageMode = hideTabLayout
                 )